- Move flow.types to main.js level - Remove comments and TODOs - Remove utils.js and place private functions in mutations and action files
